+++ src/base/behaviortree/bt_noderepeat.cpp
@@ -59,7 +59,7 @@ void BT_NodeRepeat::deserialize( QVariantMap in )
 
 	auto vcl  = in.value( "Childs" ).toList();
 	int index = 0;
-	if ( vcl.size() == m_children.size() )
+	if ( size_t( vcl.size() ) == m_children.size() )
 	{
 
 		for ( auto child : m_children )
@@ -71,7 +71,7 @@ void BT_NodeRepeat::deserialize( QVariantMap in )
 	{
 		//tree changed between saving and loading, this will have undetermined results
 		// TODO throw exception or make config option to allow or deny loading this
-		if ( vcl.size() < m_children.size() )
+		if ( size_t( vcl.size() ) < m_children.size() )
 		{
 			for ( auto vcm : vcl )
 			{
